Extract VaultProxyStatusCards, TxStatus cards, fix imports
