+++ blockchain/calls/jug.ts
@@ -1,7 +1,7 @@
 import BigNumber from 'bignumber.js'
+import { RAY, SECONDS_PER_YEAR } from 'components/constants'
 import Web3 from 'web3'
 
-import { RAY, SECONDS_PER_YEAR } from '../../components/constants'
 import { McdJug } from '../../types/web3-v1-contracts/mcd-jug'
 import { CallDef } from './callsHelpers'
