+++ blockchain/utils.ts
@@ -1,9 +1,9 @@
 import { BigNumber } from 'bignumber.js'
+import { RAD, RAY, WAD } from 'components/constants'
 import padEnd from 'lodash/padEnd'
 //@ts-ignore
 import ethAbi from 'web3-eth-abi'
 
-import { RAD, RAY, WAD } from '../components/constants'
 import { getToken } from './tokensMetadata'
 
 export function amountFromRay(amount: BigNumber): BigNumber {
