+++ features/manageVault/ManageVaultConfirmation.tsx
@@ -1,8 +1,5 @@
 import { Details } from 'components/forms/Details'
-import {
-  TxStatusCardProgress,
-  TxStatusCardSuccess,
-} from 'features/openVault/components/TxStatusCard'
+import { TxStatusCardProgress, TxStatusCardSuccess } from 'components/vault/TxStatusCard'
 import { formatAmount, formatCryptoBalance, formatPercent } from 'helpers/formatters/format'
 import { zero } from 'helpers/zero'
 import { useTranslation } from 'next-i18next'
