+++ public/locales/en/common.json
@@ -383,7 +383,7 @@
   "unlimited-allowance": "Unlimited Allowance",
   "vault": {
     "current-price": "Current {{token}}/USD Price in {{time}}",
-    "header": "{{ilk}} Vault #{{id}}",
+    "header": "{{ilk}} Vault {{id}}",
     "next-price": "Next price in {{count}} minute",
     "next-price-any-time": "Next price in <2 minutes",
     "next-price_plural": "Next price in {{count}} minutes",
@@ -456,7 +456,6 @@
   "view-on-etherscan": "View on Etherscan",
   "waiting-approval": "Waiting for approval",
   "waiting-confirmation": "Waiting for transaction confirmation",
-  "waiting-proxy-deployment": "Proxy deployment confirming",
   "proxy-deployment-confirming": "{{proxyConfirmations}} of {{safeConfirmations}}: Proxy deployment confirmed",
   "withdraw": "Withdraw",
   "withdrawn": "Withdrawn",
