+++ public/locales/pt/common.json
@@ -322,7 +322,6 @@
   },
   "waiting-approval": "Aguardando aprovação",
   "waiting-confirmation": "Aguardando confirmação da transação",
-  "waiting-proxy-deployment": "Aguardando Deploy do Proxy",
   "withdraw": "Retirar",
   "withdrawn": "Retirado",
   "wyre-badge-country": "EUA & Europa",
